[v2.2.0] Major Refactor & Stats Recording Fixed
- Split all player stats recording to its own Cog for better organization.
- Stats now record when the time elapsed has paused (and server has more than 1 player), which should be a better indicator of a game end, and will make sure to capture true end of game stats (including top player). When the time resets, this will indicate a new game has started.
- Server Status and Player Stats are polled at two different rates now. Player stats are queried faster to obtain true end game stats. Status is acquired from this data at a slower rate to not exceed Discord's message edit rate limit.
- API querying, query data, and the Inflect object have all been moved to the bot subclass to be global.
- Moved the status text channel object to be a Cog instance variable so it doesn't have to be found every loop. Also added an attempt to handle an error with editing the message when the Discord API goes down.
- Changed `/stats setstatus` to `/server setstatus`
- Moved all common strings and string functions to CommonStrings, which the Cogs now utilize.
- Re-orginized config file.
