+++ .gitignore
@@ -1,14 +1,8 @@
-# These are some examples of commonly ignored file patterns.
-# You should customize this list as applicable to your project.
-# Learn more about .gitignore:
-#     https://www.atlassian.com/git/tutorials/saving-changes/gitignore
-
 # Local config file
 config.cfg
 
-# Local music files
-audio/*
-!audio/*.txt
+# Local dev files
+TODO.txt
 
 # Python compiled bytecode and cache
 **/__pycache__/
