+++ config-example.cfg
@@ -8,12 +8,18 @@
         "Pass": "PASSWORD_HERE",
         "DB_Name": "DATABASE_NAME_HERE"
     },
+    "API": {
+        "RootURL": "https://stats.bf2mc.net",
+        "ApiURL": "https://stats.bf2mc.net/api/servers/"
+    },
     "ServerStatus": {
         "StatusVoiceChannelID": 012345678901234567,
         "ServerStatsTextChannelID": 012345678901234567,
-        "PlayerStatsTextChannelID": 012345678901234567,
         "AnnouncementTextChannelID": 012345678901234567,
-        "UpdateIntervalMinutes": 1.5,
         "OfficialIDs": [1, 3]
+    }
+    "PlayerStats": {
+        "PlayerStatsTextChannelID": 012345678901234567,
+        "QueryIntervalSeconds": 10
     }
 }
