+++ main.py
@@ -1,7 +1,7 @@
 """main.py
 
 Main file to start Backstab
-Date: 06/05/2023
+Date: 06/17/2023
 Authors: David Wolfe (Red-Thirten)
 Licensed under GNU GPLv3 - See LICENSE for more details.
 """
@@ -13,9 +13,10 @@ from discord.ext import commands
 from src import BackstabBot
 
 def main():
-    VERSION = "2.1.6"
+    VERSION = "2.2.0"
     AUTHORS = "Red-Thirten"
     COGS_LIST = [
+        "CogPlayerStats",
         "CogServerStatus"
     ]
